+++ src/shared.py
